+++ lib/loops_sdk/transactional.rb
@@ -7,32 +7,34 @@ module LoopsSdk
         make_request(method: :get, path: "v1/transactional-emails", params: { perPage: perPage, cursor: cursor })
       end
 
-      def create(name:)
-        make_request(method: :post, path: "v1/transactional-emails", body: { name: name })
+      def create(name:, transactional_group_id: nil)
+        body = { name: name, transactionalGroupId: transactional_group_id }.compact
+        make_request(method: :post, path: "v1/transactional-emails", body: body)
       end
 
-      def get(transactional_id:)
-        make_request(method: :get, path: "v1/transactional-emails/#{transactional_id}")
+      def get(id:)
+        make_request(method: :get, path: "v1/transactional-emails/#{id}")
       end
 
-      def update(transactional_id:, name:)
-        make_request(method: :post, path: "v1/transactional-emails/#{transactional_id}", body: { name: name })
+      def update(id:, name: nil, transactional_group_id: nil)
+        body = { name: name, transactionalGroupId: transactional_group_id }.compact
+        make_request(method: :post, path: "v1/transactional-emails/#{id}", body: body)
       end
 
-      def ensure_draft(transactional_id:)
-        make_request(method: :post, path: "v1/transactional-emails/#{transactional_id}/draft")
+      def ensure_draft(id:)
+        make_request(method: :post, path: "v1/transactional-emails/#{id}/draft")
       end
 
-      def publish(transactional_id:)
-        make_request(method: :post, path: "v1/transactional-emails/#{transactional_id}/publish")
+      def publish(id:)
+        make_request(method: :post, path: "v1/transactional-emails/#{id}/publish")
       end
 
-      def send(transactional_id:, email:, add_to_audience: false, data_variables: {}, attachments: [], headers: {})
+      def send(id:, email:, add_to_audience: false, data_variables: {}, attachments: [], headers: {})
         attachments = attachments.map do |attachment|
           attachment.transform_keys { |key| key == :content_type ? :contentType : key }
         end
         email_data = {
-          transactionalId: transactional_id,
+          transactionalId: id,
           email: email,
           addToAudience: add_to_audience,
           dataVariables: data_variables,