+++ spec/loops_sdk/workflows_spec.rb
@@ -1,0 +1,39 @@
+# frozen_string_literal: true
+
+require "spec_helper"
+
+RSpec.describe LoopsSdk::Workflows do
+  let(:connection) { instance_double(Faraday::Connection) }
+  let(:response) { instance_double(Faraday::Response) }
+  let(:default_headers) do
+    {
+      "Authorization" => "Bearer test-key",
+      "Content-Type" => "application/json"
+    }
+  end
+
+  before do
+    allow(LoopsSdk.configuration).to receive(:connection).and_return(connection)
+    allow(connection).to receive(:headers).and_return(default_headers)
+  end
+
+  describe ".get_node" do
+    it "makes a GET request to fetch a workflow node" do
+      expect(connection).to receive(:send).with(:get) do |&block|
+        req = double("req")
+        expect(req).to receive(:url).with("v1/workflows/workflow_123/nodes/node_456")
+        expect(req).to receive(:headers=).with(default_headers)
+        expect(req).to receive(:params=).with({})
+        expect(req).to receive(:body=).with(nil)
+        block.call(req)
+        response
+      end
+
+      allow(response).to receive(:status).and_return(200)
+      allow(response).to receive(:body).and_return('{"id":"node_456"}')
+
+      result = described_class.get_node(id: "workflow_123", node_id: "node_456")
+      expect(result).to eq({ "id" => "node_456" })
+    end
+  end
+end